Update layouts.py and notebooks
Add new function to reduce point sources by block splitting the
observation points and adding a single point source per populated block.
The coordinates of each single point is computed by reducing the
coordinates of the data points that belongs to the block. The upward
component is set in similar way as the default EQLHarmonic does: the
relative depth is proportional to the median distance to the k nearest
source points. The main difference is that the nearest neighbours are
computed on the reduced source points and not on data points. Also, add
the possibility to add a static shift.