Remove simulated noise
The addition of Gaussian noise was already commented out. I don't think
we need to add noise to the observation data and then comparing the
predictions with a target data without it. If added, then the plotted
prediction errors will be a conjunction of the prediction errors plus
the observation errors.