Remove synthetic surveys: use the Harmonica ones
The functions that creates the synthetic surveys have been moved to
Harmonica, therefore there's no need to define them inside this
repository. Removed the synthetic_surveys.py file and changed the
notebooks to use the Harmonica synthetic surveys. Had to project the
surveys inside the notebooks because they are now given in geodetic
coordinates.