Extend parameters when comparing source distributions (#21)
The set of parameters defined on the notebooks comparing source
distributions have been extended so no best prediction is achieved with
one boundary value. The only exceptions are when the parameters are only
positive (and zero) and the best prediction is achieved at one (or
zero). Gridding the airborne with a depth_shift equal to zero raised
a division by zero, so this specific value is omitted to avoid the
error.