Refactor notebooks (#25)
Split the generation of the synthetic surveys and target grid from the gridding
process and comparison. This allows to reuse synthetic surveys and target grids
on other notebooks without the need to recompute them and repeat code. Also
this ensures we are always using the same surveys and target grids. Remove old
gridding notebooks. Remove old synthetic surveys notebook.