Add run target on Makefile
Add a new run target on Makefile to run every .py file inside
`notebooks` through jupytext. I attempted to create a target-requirement
strategy, but due to jupytext updates the timestamp of every notebook it
runs, all the results will always be older than the notebooks that
creates them, so there is no point to do so. A single target to run
every notebook seemed to achieve the same goal with much simpler
structure.