Generate Jacobian only inside the window
Instead of building a Jacobian matrix for sources inside the window and
every data point, it's better to create the Jacobian for both sources
and data points inside the window. Then the residue can get updated
through predict_numba. By doing this we recompute the Jacobian matrix of
the window on each iteration, but we prevent storing a larger one into
memory.