Shorten and summarize basic EQL theory in methods
I was a bit unsure about the explanations for the theory behind the
equivalent layer that we were presenting. It felt a bit too short with
some important parts missing about the upward continuation equation and
Green's identities. But at the same time it was feeling too long for a
paper (considering we already have to explain the gradient boosting
scheme). So I figured it might be good to cheat and start from the
Cordell 1992 and skip the deductions of why we can use 1/r. That's been
done in books and we have the citation for Cordell so we don't have to
repeat what's on there.