Start a revision of the inversion section
Make some tweaks to the notation, including using B and m instead of Â
and ĉ since ^ is often used to denote an estimate. Removed some of the
^2 from the lambdas in the least-squares solutions. Removed the part
about Cholesky decomposition because that is used to solve the linear
system, not minimize the goal function (and not always, it depends on
what sklearn does). Added an ^o to the data in the equations to
distinguish between predicted and observed data.