Revise the least squares solution section
Removed the underdetermined solution because in practice we don't really
use it. Even for grid sources, our code runs the overdetermined version.
Also left the system instead of using the inverse because that is also
what we do in practice (and I hate having to always say that in practice
we don't calculate inverses).