Change references style for manuscript (#116)
Use a custom `apalike` references style: adds the doi and urls of every
publications with hyperlinks and also adds the ISBN of some publications.
Remove the urls from the `references.bib`. Fix titles in `references.bib` that
contains capital letters. Also configure fontsize and geometry of manuscript to
look better.