Add schematic figure of gradient boosting (#123)
Add a sketch that symbolizes the gradient-boosting process using the synthetic
ground survey. Add reference to the figure on the text. Renames the notebook
that creates the schematic figures. Only the figure showing the smaller sets of
sources is created through the notebook. the final figure was created through
Inkscape by adding the observation points and the location of the original
sources.