Add a section on collocation (#160)
Included a paragraph on collocation explaining what it's use for and
that it suffers from the same computational limitations of EQL. Also
broke apart the paragraph where we compare all the different EQL papers.
It was huge and a bit difficult to follow. Had to add some text to the
location strategies part otherwise it would be a single sentence.