Review the results once more (#162)
* Clarify the part about the A_k Jacobian matrices

Break apart the item that describes them to highlight the definition.

* Break up the optimization used in EQL GB

Using only data inside the window and forward modelling by summation
were locked inside a large paragraph. Use an enumerate to make it clear
that there are two optimizations and make them jump out in the text a
bit more.

* Review the source layouts results section

* Slight tweaks to window size results

* Slight tweaks to caption and section for window overlap

* Simplify caption of the synthetic data

* Tweaks to figures

Use RbBy for synthetic data since it's diverging (and seismic is a bit
too saturated at the ends). Make the random/sequential comparison single
column by plotting convergence on the same plot. Add xlabel to the
histogram in australia residuals.

* Revise the new gb random/sequential section

* Revise the australia results

* Change synthetic map to viridis and make CV figure 1 column